Remove evaluation code
