+++ Makefile
@@ -1,5 +1,5 @@
 black: FORCE
-	black scratch checkmate experiments setup.py --line-length 127
+	black scripts checkmate setup.py --line-length 127
 
 test: FORCE
 	pytest tests
