--- checkmate/__init__.py
@@ -1,2 +1,0 @@
-from .tf2.load_keras_model import get_keras_model
-from .tf2.wrapper import compile_tf2
