+++ checkmate/plot/definitions.py
@@ -1,3 +1,6 @@
+import pathlib
+from pathlib import Path
+
 PLATFORM_CHOICES = ["p32xlarge", "p32xlarge_fp16", "p2xlarge", "c524xlarge", "flops"]
 
 
@@ -18,3 +21,18 @@ def platform_memory(platform: str):
     if platform in mapping:
         return mapping[platform]
     return platform
+
+
+def checkmate_root_dir() -> Path:
+    """Returns project root folder."""
+    return Path(__file__).parent.parent.parent
+
+
+def checkmate_data_dir() -> Path:
+    """Returns project root folder."""
+    return checkmate_root_dir() / "data"
+
+
+def checkmate_cache_dir() -> Path:
+    """Returns cache dir"""
+    return pathlib.Path("/tmp") / "remat_cache"
