+++ checkmate/tf2/__init__.py
@@ -1,0 +1,2 @@
+from .wrapper import compile_tf2 as compile
+from .util.load_keras_model import get_keras_model
