+++ checkmate/tf2/util/execution_utils.py
