+++ setup.py
@@ -7,23 +7,13 @@ setup(
     packages=["checkmate"],  # find_packages()
     python_requires=">=3.5",
     install_requires=[
+        # "keras_segmentation @ https://github.com/ajayjain/image-segmentation-keras/archive/master.zip#egg=keras_segmentation-0.2.0remat",
         "matplotlib",  # this is only used once in the core checkmate package
+        "graphviz",
         "numpy",
         "pandas",
         "toposort",
         "psutil",
     ],
-    extras_require={
-        "eval": [
-            "graphviz",
-            "keras_segmentation @ https://github.com/ajayjain/image-segmentation-keras/archive/master.zip#egg=keras_segmentation-0.2.0remat",
-            "python-dotenv",
-            "ray>=0.7.5",
-            "redis",
-            "scipy",
-            "seaborn",
-            "tqdm",
-        ],
-        "test": ["graphviz", "pytest", "tensorflow>=2.0.0"],
-    },
+    extras_require={"test": ["pytest", "tensorflow>=2.0.0"]},
 )
