+++ tests/test_linear.py
@@ -2,7 +2,7 @@ import logging
 
 import numpy as np
 
-from experiments.common.graph_plotting import plot_schedule
+from checkmate.plot.graph_plotting import plot_schedule
 from checkmate.core.graph_builder import gen_linear_graph
 from checkmate.core.solvers.strategy_checkpoint_all import solve_checkpoint_all, solve_checkpoint_all_ap
 from checkmate.core.solvers.strategy_checkpoint_last import solve_checkpoint_last_node
@@ -30,7 +30,9 @@ def test_checkpoint_last():
         scheduler_result = solve_checkpoint_last_node(g)
         assert scheduler_result.feasible
         if SAVE_DEBUG_PLOTS:
-            plot_schedule(scheduler_result, False, save_file="/tmp/test_checkmate/plot_checkmate_last/{}.png".format(graph_length))
+            plot_schedule(
+                scheduler_result, False, save_file="/tmp/test_checkmate/plot_checkmate_last/{}.png".format(graph_length)
+            )
 
 
 def test_checkpoint_all_ap():
@@ -65,7 +67,10 @@ def test_chen_greedy():
             for budget in np.arange(0, 1, 0.1):
                 scheduler_result = solve_chen_greedy(g, total_cost * budget, False)
                 if scheduler_result.feasible:
-                    plot_schedule(scheduler_result, save_file="/tmp/test_checkmate/plot_chen_greedy/{}_{}.png".format(graph_length, budget))
+                    plot_schedule(
+                        scheduler_result,
+                        save_file="/tmp/test_checkmate/plot_chen_greedy/{}_{}.png".format(graph_length, budget),
+                    )
 
 
 def test_chen_greedy_ap():
@@ -79,7 +84,10 @@ def test_chen_greedy_ap():
             for budget in np.arange(0, 1, 0.1):
                 scheduler_result = solve_chen_greedy(g, total_cost * budget, False)
                 if scheduler_result.feasible:
-                    plot_schedule(scheduler_result, save_file="/tmp/test_checkmate/plot_chen_greedy_ap/{}_{}.png".format(graph_length, budget))
+                    plot_schedule(
+                        scheduler_result,
+                        save_file="/tmp/test_checkmate/plot_chen_greedy_ap/{}_{}.png".format(graph_length, budget),
+                    )
 
 
 def test_ilp():
@@ -90,6 +98,7 @@ def test_ilp():
         logging.warning("Continuing with tests, gurobi not installed")
         return
     from checkmate.core.solvers.strategy_optimal_ilp import solve_ilp_gurobi
+
     for graph_length in test_points:
         g = gen_linear_graph(graph_length)
         assert g.size == 2 * graph_length + 1
@@ -98,9 +107,13 @@ def test_ilp():
         assert scheduler_result.feasible
         if SAVE_DEBUG_PLOTS:
             for budget in np.arange(0, 1, 0.25):
-                scheduler_result = solve_ilp_gurobi(g, total_cost * budget, print_to_console=False, write_log_file=None, time_limit=15)
+                scheduler_result = solve_ilp_gurobi(
+                    g, total_cost * budget, print_to_console=False, write_log_file=None, time_limit=15
+                )
                 if scheduler_result.feasible:
-                    plot_schedule(scheduler_result, save_file="/tmp/test_checkmate/plot_ilp/{}_{}.png".format(graph_length, budget))
+                    plot_schedule(
+                        scheduler_result, save_file="/tmp/test_checkmate/plot_ilp/{}_{}.png".format(graph_length, budget)
+                    )
 
 
 # NOTE: Griewank test is disabled as the solver does not support nonlinear graphs.
