+++ tests/test_tf2_extraction.py
@@ -1,7 +1,7 @@
 import tensorflow as tf
 
 from checkmate.tf2.extraction import dfgraph_from_tf_function
-from checkmate.tf2.load_keras_model import get_keras_model
+from checkmate.tf2.util.load_keras_model import get_keras_model
 
 
 def test_mlpblock_extract():
